Improve assertions

+++ Testbed/src/main.cpp
@@ -97,7 +97,7 @@ int main()
         ImGui::SliderFloat("R", &renderer->red, 0.0, 1.0f);
         ImGui::SliderFloat("G", &renderer->green, 0.0, 1.0f);
         ImGui::SliderFloat("B", &renderer->blue, 0.0, 1.0f);
-        ImGui::SliderFloat("Rotate", &renderer->rotate, 0.0, glm::radians(90.0));
+        ImGui::SliderFloat("Rotate", &renderer->rotate, 0.0, (f32)glm::radians(90.0));
         ImGui::Text("%.3f ms/frame (%.1f FPS)", 1000.0f / ImGui::GetIO().Framerate, ImGui::GetIO().Framerate);
         ImGui::End();
         ImGui::Render();
